test_requirements.txt
tox
djangocms-admin-style
selenium